+++ inc/bouncer-instance-standalone.php
@@ -1,16 +1,20 @@
 <?php
 
 use CrowdSecBouncer\Bouncer;
+use CrowdSecBouncer\Fixes\Memcached\TagAwareAdapter as MemcachedTagAwareAdapter;
 use CrowdSecBouncer\BouncerException;
 use CrowdSecBouncer\Constants;
 use Monolog\Formatter\LineFormatter;
 use Monolog\Handler\RotatingFileHandler;
 use Monolog\Logger;
-use Symfony\Component\Cache\Adapter\AbstractAdapter;
 use Symfony\Component\Cache\Adapter\MemcachedAdapter;
 use Symfony\Component\Cache\Adapter\PhpFilesAdapter;
 use Symfony\Component\Cache\Adapter\RedisAdapter;
 use Symfony\Component\Cache\Exception\InvalidArgumentException;
+use Symfony\Component\Cache\Adapter\TagAwareAdapter;
+use Symfony\Component\Cache\Adapter\RedisTagAwareAdapter;
+use \Symfony\Component\Cache\Adapter\TagAwareAdapterInterface;
+
 
 /** @var Logger|null */
 $crowdSecLogger = null;
@@ -42,11 +46,11 @@ function getStandaloneCrowdSecLoggerInstance(string $crowdsecLogPath, bool $debu
     return $logger;
 }
 
-/** @var AbstractAdapter|null */
+/** @var TagAwareAdapterInterface|null */
 $crowdSecCacheAdapterInstance = null;
 
 function getCacheAdapterInstanceStandalone(string $cacheSystem, string $memcachedDsn, string $redisDsn, string
-$fsCachePath, bool $forcedReload = false): AbstractAdapter
+$fsCachePath, bool $forcedReload = false): TagAwareAdapterInterface
 {
     // Singleton for this function
 
@@ -57,7 +61,7 @@ $fsCachePath, bool $forcedReload = false): AbstractAdapter
 
     switch ($cacheSystem) {
         case Constants::CACHE_SYSTEM_PHPFS:
-            $crowdSecCacheAdapterInstance = new PhpFilesAdapter('', 0, $fsCachePath);
+            $crowdSecCacheAdapterInstance = new TagAwareAdapter(new PhpFilesAdapter('', 0, $fsCachePath));
             break;
 
         case Constants::CACHE_SYSTEM_MEMCACHED:
@@ -66,7 +70,9 @@ $fsCachePath, bool $forcedReload = false): AbstractAdapter
                 ' Please set a Memcached DSN or select another cache technology.');
             }
 
-            $crowdSecCacheAdapterInstance = new MemcachedAdapter(MemcachedAdapter::createConnection($memcachedDsn));
+            $crowdSecCacheAdapterInstance = new MemcachedTagAwareAdapter(
+                new MemcachedAdapter(MemcachedAdapter::createConnection($memcachedDsn))
+            );
             break;
 
         case Constants::CACHE_SYSTEM_REDIS:
@@ -76,7 +82,7 @@ $fsCachePath, bool $forcedReload = false): AbstractAdapter
             }
 
             try {
-                $crowdSecCacheAdapterInstance = new RedisAdapter(RedisAdapter::createConnection($redisDsn));
+                $crowdSecCacheAdapterInstance = new RedisTagAwareAdapter(RedisAdapter::createConnection($redisDsn));
             } catch (InvalidArgumentException $e) {
                 throw new BouncerException('Error when connecting to Redis.'.
                 ' Please fix the Redis DSN or select another cache technology.');